move engine logic into engineview
