+++ Modules/@babylonjs/react-native/windows/BabylonReactNative/EngineView.idl
@@ -1,0 +1,9 @@
+namespace BabylonReactNative
+{
+    [webhosthidden]
+    [default_interface]
+    runtimeclass EngineView :
+        Windows.UI.Xaml.Controls.SwapChainPanel {
+      EngineView();
+    };
+} // namespace BabylonReactNative
