+++ Modules/@babylonjs/react-native/windows/BabylonReactNative/EngineViewManager.h
@@ -1,8 +1,7 @@
 #pragma once
 #include "NativeModules.h"
 #include "winrt/Microsoft.ReactNative.h"
-#include "JSI/jsi.h"
-#include <unordered_set>
+#include "EngineView.h"
 
 namespace winrt::BabylonReactNative::implementation {
 
@@ -37,26 +36,8 @@ namespace winrt::BabylonReactNative::implementation {
         winrt::Microsoft::ReactNative::ConstantProviderDelegate ExportedCustomDirectEventTypeConstants() noexcept;
 
     private:
-        void OnSizeChanged(winrt::Windows::Foundation::IInspectable const& sender, winrt::Windows::UI::Xaml::SizeChangedEventArgs const& args);
-        void OnPointerPressed(winrt::Windows::Foundation::IInspectable const& sender, winrt::Windows::UI::Xaml::Input::PointerRoutedEventArgs const& args);
-        void OnPointerMoved(winrt::Windows::Foundation::IInspectable const& sender, winrt::Windows::UI::Xaml::Input::PointerRoutedEventArgs const& args);
-        void OnPointerReleased(winrt::Windows::Foundation::IInspectable const& sender, winrt::Windows::UI::Xaml::Input::PointerRoutedEventArgs const& args);
-        void OnRendering();
-
         winrt::Microsoft::ReactNative::IReactContext _reactContext{ nullptr };
-        winrt::Windows::UI::Xaml::Controls::SwapChainPanel _swapChainPanel{ nullptr };
-        size_t _swapChainPanelWidth{ 1 };
-        size_t _swapChainPanelHeight{ 1 };
-
-        struct RevokerData
-        {
-            winrt::Windows::UI::Xaml::FrameworkElement::SizeChanged_revoker SizeChangedRevoker{};
-            winrt::Windows::UI::Xaml::FrameworkElement::PointerPressed_revoker PointerPressedRevoker{};
-            winrt::Windows::UI::Xaml::FrameworkElement::PointerMoved_revoker PointerMovedRevoker{};
-            winrt::Windows::UI::Xaml::FrameworkElement::PointerReleased_revoker PointerReleasedRevoker{};
-            winrt::Windows::UI::Xaml::Media::CompositionTarget::Rendering_revoker RenderingRevoker{};
-        };
-        RevokerData _revokerData{};
+        winrt::BabylonReactNative::EngineView _engineView{ nullptr };
     };
 
 } // namespace winrt::BabylonReactNative::implementation
